add user setting fullness message 5

+++ src/communicate/main.py
@@ -1639,7 +1639,7 @@ def compose_msg_on_user_setting_fullness(cur, user_id: int) -> Union[str, None]:
             pref_urgency, pref_notif_type, pref_region_old, pref_forum = raw_data
 
         list_of_settings = [pref_notif_type, pref_region_old, pref_coords, pref_radius, pref_age, pref_forum]
-        user_score = round(sum(list_of_settings)/len(list_of_settings)*100, 0)
+        user_score = int(round(sum(list_of_settings)/len(list_of_settings)*100, 0))
 
         logging.info(f'{list_of_settings}')
         logging.info(f'{user_score}')
